requirements.txt
# Core dependencies
chromadb==0.5.0
click==8.1.7
python-dotenv==1.0.1
fastapi==0.111.0
uvicorn==0.30.1
PyGithub==2.3.0
pydantic==2.8.2
rich==13.7.1
requests==2.32.3
